Experimental untested SQLite migration
This commit marks the initial migration of the entire plugin to use SQLite instead of YAML.
SQLite databases are easier to maintain and update in cases of major plugin changes and is required in order for the plugin to have backward compatibility in future updates.
This update makes any earlier builds of v5.0-Early-Alpha incompatible with this version.